==> notebooks/Script 1 - Responses Cleaning.ipynb <==
# Dataset Cleaning
This dataset is my own collected data from surveys.

Since the table width is not too large, I decided to clean up the columns via Excel. 

The main objectives of this notebook are:
- Observe the contents of the dataset,
- melt the columns for characters and movies,
- separate data with proper separator from those with improper formatting,
- remove empty entries,
- export the cleaned data.

import pandas as pd
import numpy as np

## Loading the dataset

df = pd.read_csv('../data/Data 1 - Raw Survey Responses.csv', encoding='unicode_escape')
df.head(-10)

## Initial exploration of the data

print("Dataset shape:", df.shape)
print("Dataset columns:", df.columns)

**Observations**
- Some columns are not necessary.

**Actions(s)**
- Eliminate "consent" and text-id columns. 
- Insert index column as id.

df_dropped = df.drop(['timestamp', 'username', 'consent', 'age_confirmation', 'comments'], axis = 1)

# Add id column before melting, then rename `index` as `id`
df_dropped.reset_index(inplace=True)
df_dropped.rename(columns = {'index':'id', 'mbti':'person_mbti'}, inplace = True)
df_dropped

## Melt for each column

df_melt = pd.melt(df_dropped, id_vars=['id', 'person_mbti', 'mbti_test'], var_name='char_entry', value_name='char_movie', col_level=None, ignore_index=True)
df_melt.drop(['char_entry'], axis=1, inplace=True)
df_melt.sort_values('id', ascending=True)
df_melt.dropna(inplace=True)
df_melt


## Split data


import re
def remove_spaces(s):
    s = " ".join(s.split())

    if s[0]==' ':
        s = re.sub(r"^\s+", "", s)
    if s[-1]==' ':
        re.sub(r"\s+$", "", s)

    return s

print(remove_spaces(' Iron Man  '))

df_movie_char_split = df_melt
new_info = df_movie_char_split['char_movie'].str.split('//', n=1, expand=True)
df_movie_char_split['character'] = new_info[0]
df_movie_char_split['movie'] = new_info[1]

df_movie_char_split.drop(columns=['char_movie'], inplace=True)
df_movie_char_split.sort_values('id', inplace=True)

df_movie_char_split['movie'].apply(remove_spaces)
df_movie_char_split['character'].apply(remove_spaces)

print(df_movie_char_split)

df_movie_char_split.to_csv('../data/Data 2 - Transformed Survey Responses.csv')
