==> notebooks/Script 4 - Char MBTI Franchise Expansion.ipynb <==
import requests
import json
import urllib.request
import pandas as pd
import ast

# Header for API
headers = {
    "accept": "application/json",
    "Authorization": "Bearer eyJhbGciOiJIUzI1NiJ9.eyJhdWQiOiJjYmMwMDBjMGE5YzZjZmNkMDYxYWIyN2EzMjA1YjM0NiIsInN1YiI6IjY2NDRlMzUzM2YwMGEzYTg5YjViZDFkMiIsInNjb3BlcyI6WyJhcGlfcmVhZCJdLCJ2ZXJzaW9uIjoxfQ._E0exiVUWU6c_FyyCeGdqQpaOM_tFDreyWlV6bmclPk"
}

characters_mbti_df = pd.read_csv('../data/Data 8 - Franchise-expanded Movies from Data 2.csv', encoding='unicode_escape')
# survey_responses_df.drop(columns='Unnamed: 0', inplace=True)
characters_mbti_df

franchise_name = "Indiana Jones (Series)"

## Getting the movie details

list_of_franchise_movies = pd.read_csv('../data/Data 7 - Raw List of Franchised Movies.csv', encoding='unicode_escape')
list_of_franchise_movies

class expanding_movies_to_cast:
    def __init__(self, input):
        self.input = input
        self.movie_list_in_franchise = None
        self.franchise_dict_list = []
        self.failed_list = []
        self.franchise_movie_id_list = []
        self.franchise_details_df = None
        self.char = None

    def append_to_franchise_movies_list(self, item):
        self.franchise_dict_list.append(item)

    def append_to_failed_list(self, item):
        self.failed_list.append(item)

    def fetch_characters_in_franchise(self): 
        franchise_characters_mbti_df = characters_mbti_df[characters_mbti_df['movie']==self.input]

        franchise_chars = list(franchise_characters_mbti_df['char'])
        chars_mbti = list(franchise_characters_mbti_df['mbti'])

        retain_chars = 'a-zA-Z0-9\s\-:–'

        chars_in_franchise = pd.DataFrame({'char': franchise_chars, 'movie': self.input, 'mbti': chars_mbti})
        chars_in_franchise['char'] = chars_in_franchise['char'].str.replace(f'[^{retain_chars}]', '', regex=True)
        return chars_in_franchise
    
    def fetch_movies_in_franchise(self):
        self.movie_list_in_franchise = list(list_of_franchise_movies[list_of_franchise_movies['franchise']==self.input]['movie'])   
        try:
            self.movie_list_in_franchise = [movie.replace('\xa0', ' ') for movie in self.movie_list_in_franchise]
        except:
            pass
        return self.movie_list_in_franchise
    
    def fetch_movie_details(self):
        franchise_dict_list = []
        failed_list = []
        for i, title in enumerate(self.movie_list_in_franchise):
            try:
                title_replace = title.replace(' ', '+')
                split_title = title_replace.split('+(', maxsplit=1)
                title_part = split_title[0]
                year_part = split_title[1].replace(')', '')
                
                url = 'https://api.themoviedb.org/3/search/movie?query=' + title_part + '&year=' + year_part + '&api_key=cbc000c0a9c6cfcd061ab27a3205b346'
            except:
                url = 'https://api.themoviedb.org/3/search/movie?query=' + title + '&api_key=cbc000c0a9c6cfcd061ab27a3205b346'

            try:
                response = requests.get(url, headers=headers)
                data = response.json()

                # While data above is a dictionary, it contains more dictionaries within it, so narrow down
                movie_data = data['results'][0]
                # print(i, movie_data['title'])
                self.append_to_franchise_movies_list(movie_data)
            except:
                self.append_to_failed_list(title)
                # print(i, 'Failed fetching ', split_title[0], len(failed_list))
        return self.franchise_dict_list, self.failed_list
    
    def franchise_movies_pipeline(self):
        franchise_details_df = pd.DataFrame(data=self.franchise_dict_list)
        franchise_details_df.drop(columns=['adult', 'backdrop_path', 'original_language', 'original_title', 'overview', 'poster_path', 'video'], inplace=True)
        self.franchise_movie_id_list = list(franchise_details_df['id'])
        return franchise_details_df
    
    def get_chars_in_movie_id_list(self):
        cast_dict_list = []
        dict_list = []
        for movie_id in self.franchise_movie_id_list:
            url = 'https://api.themoviedb.org/3/movie/' + str(movie_id) + '/credits?language=en-US'
            response = requests.get(url, headers=headers)
            data = response.json()
            cast_len = len(data['cast'])
            movie_id_dict = {'movie_id': movie_id}
            for cast in range(cast_len):
                new_cast = data['cast'][cast]
                new_cast.update(movie_id_dict)
                cast_dict_list.append(data['cast'][cast])
        franchise_cast_df = pd.DataFrame(data=cast_dict_list)
        franchise_cast_drop_df = franchise_cast_df.drop(columns=['adult', 'original_name', 'profile_path', 'credit_id', 'order', 'cast_id'])
        return franchise_cast_drop_df

    def franchise_cast_merge_pipeline(self):
        franchise_cast_df = self.get_chars_in_movie_id_list()
        franchise_details_df = self.franchise_movies_pipeline()
        franchise_cast_details_merged = franchise_details_df.merge(franchise_cast_df, how='outer', left_on='id', right_on='movie_id')
        franchise_cast_details_dropped = franchise_cast_details_merged[franchise_cast_details_merged['known_for_department']=='Acting']
        return franchise_cast_details_dropped

franchise = expanding_movies_to_cast(franchise_name)
franchise_df = franchise.fetch_characters_in_franchise()
franchise.fetch_movies_in_franchise()
franchise.fetch_movie_details()
franchise_details = franchise.franchise_movies_pipeline()
franchise.get_chars_in_movie_id_list()
franchise_cast_merged = franchise.franchise_cast_merge_pipeline()

franchise_df

franchise_details

franchise_cast_merged

## Getting the characters in each movie

def get_first_word(s):
    # Strip leading/trailing whitespace and split the string
    words = s.strip().split()
    # Return the first word if the list is not empty, else return an empty string
    return words[0] if words else ''
# get_first_word('Loki Laufeyson')

def get_first_two_words(s):
    # Strip leading/trailing whitespace and split the string
    words = s.strip().split()
    # Join and return the first two words if the list is not empty, else return an empty string
    return ' '.join(words[:2])

def find_movies(character_name):
    """
    Input: String of character name.
    Output: List of movies where the character is present in.
    """
    bool_series = franchise_cast_merged['character'].str.contains(character_name, na=False)
    if bool_series.any():
        index = bool_series[bool_series].index
        movies_per_char_df = franchise_cast_merged.loc[index]
        movies_per_char_df.drop_duplicates('id_x', inplace=True)
        movies_list_of_char = list(movies_per_char_df['title'])
        return movies_list_of_char
    return franchise_name

def movies_per_character(char):
    '''
    Input: String of character name.
    Output: List of movies where the character is present in.
    '''
    try:
        two_words = get_first_two_words(char)
        movies_list_of_char = find_movies(two_words)
    except:
        movies_list_of_char = franchise_name

    if not movies_list_of_char:
        try:
            first_word = get_first_word(char)
            movies_list_of_char = find_movies(first_word)
        except:
            movies_list_of_char = franchise_name

    return movies_list_of_char


franchise_list_movies_df = franchise_df.copy()

for i in range(len(franchise_list_movies_df)):
    franchise_list_movies_df.loc[i, "movie"] = movies_per_character(str(franchise_list_movies_df.at[i, 'char']))

franchise_list_movies_df

new_characters_mbti_df = characters_mbti_df.copy()
retain_chars = 'a-zA-Z0-9\s\-'
new_characters_mbti_df['char'] = new_characters_mbti_df['char'].str.replace(f'[^{retain_chars}]', '', regex=True)
new_characters_mbti_df.set_index('char', inplace=True)
new_characters_mbti_df

new_franchise_list_movies_df = franchise_list_movies_df
new_franchise_list_movies_df.set_index('char', inplace=True)
new_franchise_list_movies_df

new_titles_merge = new_characters_mbti_df.copy() 

for i in range(len(new_franchise_list_movies_df)):
    char = new_franchise_list_movies_df.index[i]
    # print(char)
    # print(new_titles_merge.loc[char])
    new_titles_merge.loc[char, 'movie'] = str(new_franchise_list_movies_df.loc[char, 'movie'])

new_titles_merge.to_csv('../data/Data 8 - Franchise-expanded Movies from Data 2.csv')
