==> notebooks/Script 5 - Responses and Char MBTI Merging.ipynb <==
import requests
import json
import urllib.request
import pandas as pd
import ast

response_details_merge = pd.read_csv('../data/Data 6 - Merged Responses and Movie Details.csv', encoding='unicode_escape')
response_details_merge.head()

# Actions
### Columns to be removed
- `Unnamed: 0`
- `movie` and `title` since `new_title` was assigned to hold the actual movie title in Script 4
- `vote_count`
### Renames
- `id_x` as `person_id`
- `new_title` as `movie`
- `id_y` as `movie_id`
- `popularity` as `movie_popularity`
- `vote_average` as `movie_vote_avg`
### Others
- Change `release_date` to only capture year information as `movie_year`


response_details = response_details_merge.copy()

response_details.drop(columns = ['person_mbti', 'movie', 'title', 'vote_count'], inplace=True)
response_details.rename(columns = {'id_x':'person_id', 'new_title':'movie', 'id_y':'movie_id', 'popularity':'movie_popularity', 'vote_average':'movie_vote_avg'}, inplace = True)
response_details['movie_year'] = response_details_merge['release_date'].astype(str).str[:4]
response_details.drop(columns = ['release_date'], inplace=True)


retain_chars = 'a-zA-Z0-9\s\-:–\(\)'

response_details['character'] = response_details['character'].str.lower().replace(f'[^{retain_chars}]', '', regex=True)
response_details['movie'] = response_details['movie'].str.lower().replace(f'[^{retain_chars}]', '', regex=True)

response_details

char_mbti_df = pd.read_csv('../data/Data 8 - Franchise-expanded Movies from Data 2.csv', encoding='unicode_escape')

char_mbti_df['char'] = char_mbti_df['char'].str.lower().replace(f'[^{retain_chars}]', '', regex=True)
char_mbti_df['movie'] = char_mbti_df['movie'].str.lower().replace(f'[^{retain_chars}]', '', regex=True)

char_mbti_df


zero = response_details.iloc[-1]
char = zero['character']
movie = zero['movie']
print(char, movie)

char_mbti_df_copy = char_mbti_df.copy()
query_bool = char_mbti_df_copy['char'].str.contains(char) & char_mbti_df_copy['movie'].str.contains(movie)
# query_bool =  query_bool[query_bool.values==True]
char_mbti_df_copy = char_mbti_df_copy[query_bool]
char_mbti_df_copy
# print(query_bool)

char_mbti_df_copy = char_mbti_df.copy()

failed_index = []
failed_char = []
failed_movie = []

k = len(response_details)
# k = 5

for i in range(k):
    zero = response_details.iloc[i]
    char = zero['character']
    movie = zero['movie']
    # print(char, movie)

    query_bool = char_mbti_df_copy['char'].str.contains(char) & char_mbti_df_copy['movie'].str.contains(movie)
    char_mbti_query = char_mbti_df_copy[query_bool]
    try:
        char_df_subset = char_mbti_query.iloc[0]
        # print(i, char_df_subset['char'], char_df_subset['mbti'])
    except:
        print(i, '      FAILED!  ', char,  '    ', movie)
        failed_index.append(i)
        failed_char.append(char)
        failed_movie.append(movie)

failed_df = pd.DataFrame({'failed_index': failed_index, 'failed_char': failed_char, 'failed_movie': failed_movie})
failed_df
