==> notebooks/Script 6 - TMDb API Fetching.ipynb <==
import requests
import json
import urllib.request
import pandas as pd
import ast

# Predefine header for TMDb API
headers = {
    "accept": "application/json",
    "Authorization": "Bearer eyJhbGciOiJIUzI1NiJ9.eyJhdWQiOiJjYmMwMDBjMGE5YzZjZmNkMDYxYWIyN2EzMjA1YjM0NiIsInN1YiI6IjY2NDRlMzUzM2YwMGEzYTg5YjViZDFkMiIsInNjb3BlcyI6WyJhcGlfcmVhZCJdLCJ2ZXJzaW9uIjoxfQ._E0exiVUWU6c_FyyCeGdqQpaOM_tFDreyWlV6bmclPk"
}

survey_responses_df = pd.read_csv('../data/Data 2 - Transformed Survey Responses.csv', encoding='unicode_escape')
survey_responses_df

# Turn movie columns into a list
movie_list = list(survey_responses_df['movie'])

movie_dict_list = []
failed_list = []

# Iterate over all the movies listed above to compare spelling inputs
for i, title in enumerate(movie_list):
    
    # Try condition is for input with year included, processing paretheses; returns url with year
    try:
        title_replace = title.replace(' ', '+')
        split_title = title_replace.split('+(', maxsplit=1)
        title_part = split_title[0]
        year_part = split_title[1].replace(')', '')
        
        url = 'https://api.themoviedb.org/3/search/movie?query=' + title_part + '&year=' + year_part + '&api_key=cbc000c0a9c6cfcd061ab27a3205b346'
    
    # Exception: no open parenthesis is found; returns url with title only
    except:
        url = 'https://api.themoviedb.org/3/search/movie?query=' + title + '&api_key=cbc000c0a9c6cfcd061ab27a3205b346'
    
    print(i, split_title[0])

    # Attempt to fetch movie title in API
    try:
        response = requests.get(url, headers=headers)
        data = response.json()
        
        # 'data' above is a nested dictionary, so narrow down to important details
        movie_data = data['results'][0]
        print(i, movie_data['title'])
        movie_dict_list.append(movie_data)

    # Else, note for checking on errors
    except:
        failed_list.append(title)
        print('Failed', len(failed_list))
    print()
    
    survey_responses_df.at[i, 'new_title'] = movie_data['title']
survey_responses_df

# Trial function for individual search, to double check on input manually. Basically the same above but only here for test cases.
# def movie_query(movie_list):
#     """
#     Arg: list movie_list: a list of movie(s).
#     Ret: displays comparison of title input and title in API.
#     """
#     extra_movie_dict_list = []
#     extra_failed_list = []
#     for i, title in enumerate(movie_list):
#         try:
#             title_replace = title.replace(' ', '+')
#             split_title = title_replace.split('+(', maxsplit=1)
#             title_part = split_title[0]
#             year_part = split_title[1].replace(')', '')
            
#             url = 'https://api.themoviedb.org/3/search/movie?query=' + title_part + '&year=' + year_part + '&api_key=cbc000c0a9c6cfcd061ab27a3205b346'
            
#         except:
#             url = 'https://api.themoviedb.org/3/search/movie?query=' + title + '&api_key=cbc000c0a9c6cfcd061ab27a3205b346'
        
#         print(i, split_title[0])

#         try:
#             response = requests.get(url, headers=headers)
#             data = response.json()
            
#             movie_data = data['results'][0]
#             print(i, movie_data['title'])
#             extra_movie_dict_list.append(movie_data)
#         except:
#             extra_failed_list.append(title)
#             print('Failed', len(extra_failed_list))
#         print()

# movie_query(['wall.e'])

movie_details_df = pd.DataFrame(data=movie_dict_list)
movie_details_df.drop(columns=['adult', 'backdrop_path', 'original_language', 'original_title', 'overview', 'poster_path', 'video'], inplace=True)
movie_details_df

## Merging survey responses and movie details

unique_movies = movie_details_df.drop_duplicates('title')
# unique_movies

movie_details_df.to_csv('../data/Data 5 - TMDb API movie details.csv')

response_details_merge = survey_responses_df.merge(unique_movies, how='left', left_on='new_title', right_on='title')

response_details_merge.to_csv('../data/Data 6 - Merged Responses and Movie Details.csv')
response_details_merge

response_details_merge = pd.read_csv('../data/Data 6 - Merged Responses and Movie Details.csv')
response_details_merge.drop(columns=['Unnamed: 0', 'mbti_test', 'movie', 'title', 'vote_count'], inplace=True)

response_details_merge.rename(columns = {'id_x':'person_id', 'new_title':'movie', 'id_y':'movie_id', 'popularity':'movie_popularity', 'vote_average':'movie_vote_avg'}, inplace = True)

# Get first 4 text-characters as movie_year
response_details_merge['movie_year'] = response_details_merge['release_date'].str[:4].astype(int)
response_details_merge.drop(columns = ['release_date'], inplace=True)

response_details_merge

## Adding MBTI

characters_mbti_df = pd.read_csv('../data/Data 8 - Franchise-expanded Movies from Data 2.csv', encoding='unicode_escape')
characters_mbti_df

import re

def remove_spaces(s:str):
    """
    Arg: s (str): input string with possible excess spaces
    Ret: s (str): string free from excess spaces
    """
    s = " ".join(s.split())
    if s[0]==' ':
        s = re.sub(r"^\s+", "", s)
    if s[-1]==' ':
        re.sub(r"\s+$", "", s)
    return s

# recorded_chars = []
# recorded_movies = []
# missing_chars = []
# missing_movies = []
# characters_mbti_df['char'] = characters_mbti_df['char'].str.lower()
# characters_mbti_df['movie'] = characters_mbti_df['movie'].str.lower()

# def char_mbti_search(char_find, movie_find):
#     bool_series = characters_mbti_df['char'].str.contains(char_find) & characters_mbti_df['movie'].str.contains(movie_find)
#     if bool_series.any():
#         index = bool_series.idxmax()  # Or use bool_series[bool_series].index for all matches
#         # print(index)
#         recorded_chars.append(char_find)
#         recorded_movies.append(movie_find)
#         return characters_mbti_df.loc[index]  
#     else:
#         # print(f"No match found for character '{char_find}' in movie '{movie_find}'")
#         missing_chars.append(char_find)
#         missing_movies.append(movie_find)
#         return None


# # char_list = list(try_df['character'])
# # movie_list = list(try_df['movie'])
# # print(char_list)

# # char_list = 'Tony'
# # movie_list = 'Iron Man'

# # print(char_mbti_search('Gazelle', 'Zootopia'))

# import re
# def remove_spaces(s):
#     s = " ".join(s.split())
#     if s[0]==' ':
#         s = re.sub(r"^\s+", "", s)
#     if s[-1]==' ':
#         re.sub(r"\s+$", "", s)
#     return s

# for i in range(len(response_details_merge)):
#     char_i = remove_spaces(str(response_details_merge['character'][i]).lower())
#     movie_i = remove_spaces(str(response_details_merge['movie'][i]).lower())

#     try:
#         movie_split = movie_i.split(' (', maxsplit=1)
#         movie_i = movie_split[0]
#     except:
#         pass
    
#     print((char_i, movie_i))
#     char_mbti_search(char_i, movie_i)


# char_list = "poulain"
# movie_list = "Amélie"

# char_list = "wall-e"
# movie_list = "WALLÂ·E"

# char_list = "Haku"
# movie_list = "Spirited Away"

# print(char_mbti_search(char_list.lower(), movie_list.lower()))

# print("found: ", len(recorded_chars))
# print("not found:", len(missing_chars))

# for i in range(len(recorded_chars)):
#     print(recorded_chars[i], "  from  ", recorded_movies[i])

# for i in range(len(missing_chars)):
#     print(missing_chars[i], "  from  ", missing_movies[i])

characters_mbti_df['char'] = characters_mbti_df['char'].str.lower()
characters_mbti_df['movie'] = characters_mbti_df['movie'].str.lower()

with_char_mbti = response_details_merge.copy()

def get_char_mbti(char:str, movie:str):
    """"
    Args: 
        char (str): movie character
        movie (str): movie title
    Returns:
        str: mbti of the character in the movie
    """
    # Get boolean based on string presence condition
    bool_series_charmbti = characters_mbti_df['char'].str.contains(char) & characters_mbti_df['movie'].str.contains(movie)

    if bool_series_charmbti.any():
        index = bool_series_charmbti.idxmax()
        return characters_mbti_df.loc[index, 'mbti']  
    else:
        return None
    
# Iterate over all the entries of chars and movies
for i in range(len(with_char_mbti)):
    char = remove_spaces(with_char_mbti.loc[i, 'character'].lower())
    movie = remove_spaces(with_char_mbti.loc[i, 'movie'].lower())

    # Directly add the mbti to the 'char_mbti' column
    with_char_mbti.loc[i, 'char_mbti'] = get_char_mbti(char, movie)

with_char_mbti_dropped = with_char_mbti[with_char_mbti['char_mbti'].notna()]
with_char_mbti_dropped

with_char_mbti_dropped.to_csv('../data/Data 9 - Final Dataset.csv')
